requirements.txt
tensorflow==1.8.0
keras==2.1.6
numpy==1.14.3
python-twitter==3.4.1
psycopg2-binary==2.7.4
nltk==3.3
